Optimize RandomApi::nextString
Avoid boxing of int (for comprehension) and char (with the scala
StringBuilder).  Use an Array instead of capacity specified StringBuilder.

Generally it's faster to use java StringBuilder instead of the scala
StringBuilder wrapper. But Arrays are even faster when you know the
length in advance.

Side note: Random.nextInt is twice as fast for powers of 2, so if we expanded
characters to 64 instead of 62, this method could be even faster.

Also:
- Clean up oneOf
- Optimize odds