Optimize TotalWrapper
Because of the use of evidence, the types for TotalWrapper
are not being elided, leading to 10-15 JVM instruction methods
for OpaqueInt::atMost.  This is especially problematic because these
methods are tagged 'inline', so they are significantly increasing
size of the methods which use OpaqueInts.

I've verified these changes now correctly elide types entirely,
but this implementation is extremely fragile. Because scala sees
evidence for conversion, it will implicitly use the evidence given
a chance, which we don't want.

The best I could come up with is a scary warning. :-\